requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
opencv-python
pillow
numpy
